backend/requirements.txt
Django==5.1.6
openai==1.61.1
python-dotenv==1.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
pyjwt==2.10.1
pytz==2025.1
psycopg2-binary==2.9.10
paddleocr==2.9.1
paddlepaddle==2.6.2
pdf2image==1.17.0